Might as well merge this

+++ include/Game/Util/JMapIdInfo.h
@@ -9,8 +9,8 @@ public:
 
     }
 
-    JMapIdInfo(s32, s32);
-    JMapIdInfo(s32, const JMapInfoIter &);
+    JMapIdInfo(long, long);
+    JMapIdInfo(long, const JMapInfoIter &);
     JMapIdInfo(const JMapIdInfo &);
 
     void initalize(s32, const JMapInfoIter &);
